Init (blockchain core, POW consensus, leveldb, api)

+++ src/core/timestamp.py
@@ -1,5 +1,5 @@
 from datetime import datetime
 
 
-def now() -> int:
+def now() -> float:
     return datetime.now().timestamp()
